client/beefy: fix voter init tests
