+++ client/beefy/src/aux_schema.rs
@@ -75,215 +75,31 @@ where
 #[cfg(test)]
 pub(crate) mod tests {
 	use super::*;
-	use crate::{
-		keystore::tests::Keyring,
-		tests::{make_beefy_ids, BeefyTestNet},
-		KnownPeers,
-	};
-	use beefy_primitives::{
-		crypto::Signature, known_payloads, Commitment, Payload, SignedCommitment,
-		VersionedFinalityProof,
-	};
-	use futures::executor::block_on;
-	use parking_lot::Mutex;
-	use sc_client_api::{Backend as BackendT, BlockchainEvents, HeaderBackend};
+	use crate::tests::BeefyTestNet;
 	use sc_network_test::TestNetFactory;
-	use std::sync::Arc;
-	use substrate_test_runtime_client::{runtime::Block, ClientExt};
 
-	fn load_persistent_helper(
-		net: &mut BeefyTestNet,
-		finality: &mut Fuse<FinalityNotifications<Block>>,
-	) -> PersistedState<Block> {
-		let backend = net.peer(0).client().as_backend();
-		let api = Arc::new(crate::tests::two_validators::TestApi {});
-		let known_peers = Arc::new(Mutex::new(KnownPeers::new()));
-		let gossip_validator =
-			Arc::new(crate::communication::gossip::GossipValidator::new(known_peers));
-		let mut gossip_engine = sc_network_gossip::GossipEngine::new(
-			net.peer(0).network_service().clone(),
-			"/beefy/whatever",
-			gossip_validator,
-			None,
-		);
-
-		block_on(load_persistent(&*backend, &*api, &mut gossip_engine, finality, 1)).unwrap()
+	// also used in tests.rs
+	pub fn verify_persisted_version<B: BlockT, BE: Backend<B>>(backend: &BE) -> bool {
+		let version: u32 = load_decode(backend, VERSION_KEY).unwrap().unwrap();
+		version == CURRENT_VERSION
 	}
 
 	#[test]
-	fn should_initialize_voter_at_genesis() {
-		let keys = &[Keyring::Alice];
-		let validator_set = ValidatorSet::new(make_beefy_ids(keys), 0).unwrap();
+	fn should_load_persistent_sanity_checks() {
 		let mut net = BeefyTestNet::new(1);
 		let backend = net.peer(0).client().as_backend();
 
-		// push 15 blocks with `AuthorityChange` digests every 10 blocks
-		net.generate_blocks_and_sync(15, 10, &validator_set, false);
+		// version not available in db -> None
+		assert_eq!(load_persistent(&*backend).unwrap(), None);
 
-		let mut finality = net.peer(0).client().as_client().finality_notification_stream().fuse();
+		// populate version in db
+		write_current_version(&*backend).unwrap();
+		// verify correct version is retrieved
+		assert_eq!(load_decode(&*backend, VERSION_KEY).unwrap(), Some(CURRENT_VERSION));
 
-		// finalize 13 without justifications
-		let hashof13 =
-			backend.blockchain().expect_block_hash_from_id(&BlockId::Number(13)).unwrap();
-		net.peer(0).client().as_client().finalize_block(hashof13, None).unwrap();
+		// version is available in db but state isn't -> None
+		assert_eq!(load_persistent(&*backend).unwrap(), None);
 
-		// load persistent state - nothing in DB, should init at session boundary
-		let persisted_state = load_persistent_helper(&mut net, &mut finality);
-
-		// Test initialization at session boundary.
-		// verify voter initialized with two sessions starting at blocks 1 and 10
-		let sessions = persisted_state.voting_oracle().sessions();
-		assert_eq!(sessions.len(), 2);
-		assert_eq!(sessions[0].session_start(), 1);
-		assert_eq!(sessions[1].session_start(), 10);
-		let rounds = persisted_state.active_round().unwrap();
-		assert_eq!(rounds.session_start(), 1);
-		assert_eq!(rounds.validator_set_id(), validator_set.id());
-
-		// verify next vote target is mandatory block 1
-		assert_eq!(persisted_state.best_beefy_block(), 0);
-		assert_eq!(persisted_state.best_grandpa_block(), 13);
-		assert_eq!(
-			persisted_state
-				.voting_oracle()
-				.voting_target(persisted_state.best_beefy_block(), 13),
-			Some(1)
-		);
-
-		// verify state also saved to db
-		let version: u32 = load_decode(&*backend, VERSION_KEY).unwrap().unwrap();
-		assert_eq!(version, CURRENT_VERSION);
-		let state = load_decode::<_, PersistedState<Block>>(&*backend, WORKER_STATE)
-			.unwrap()
-			.unwrap();
-		assert_eq!(state, persisted_state);
-	}
-
-	#[test]
-	fn should_initialize_voter_when_last_final_is_session_boundary() {
-		let keys = &[Keyring::Alice];
-		let validator_set = ValidatorSet::new(make_beefy_ids(keys), 0).unwrap();
-		let mut net = BeefyTestNet::new(1);
-		let backend = net.peer(0).client().as_backend();
-
-		// push 15 blocks with `AuthorityChange` digests every 10 blocks
-		net.generate_blocks_and_sync(15, 10, &validator_set, false);
-
-		let mut finality = net.peer(0).client().as_client().finality_notification_stream().fuse();
-
-		// finalize 13 without justifications
-		let hashof13 =
-			backend.blockchain().expect_block_hash_from_id(&BlockId::Number(13)).unwrap();
-		net.peer(0).client().as_client().finalize_block(hashof13, None).unwrap();
-
-		// import/append BEEFY justification for session boundary block 10
-		let commitment = Commitment {
-			payload: Payload::from_single_entry(known_payloads::MMR_ROOT_ID, vec![]),
-			block_number: 10,
-			validator_set_id: validator_set.id(),
-		};
-		let justif = VersionedFinalityProof::<_, Signature>::V1(SignedCommitment {
-			commitment,
-			signatures: vec![None],
-		});
-		let hashof10 =
-			backend.blockchain().expect_block_hash_from_id(&BlockId::Number(10)).unwrap();
-		backend
-			.append_justification(hashof10, (BEEFY_ENGINE_ID, justif.encode()))
-			.unwrap();
-
-		// Test corner-case where session boundary == last beefy finalized,
-		// expect rounds initialized at last beefy finalized 10.
-
-		// load persistent state - nothing in DB, should init at session boundary
-		let persisted_state = load_persistent_helper(&mut net, &mut finality);
-
-		// verify voter initialized with single session starting at block 10
-		assert_eq!(persisted_state.voting_oracle().sessions().len(), 1);
-		let rounds = persisted_state.active_round().unwrap();
-		assert_eq!(rounds.session_start(), 10);
-		assert_eq!(rounds.validator_set_id(), validator_set.id());
-
-		// verify block 10 is correctly marked as finalized
-		assert_eq!(persisted_state.best_beefy_block(), 10);
-		assert_eq!(persisted_state.best_grandpa_block(), 13);
-		// verify next vote target is diff-power-of-two block 12
-		assert_eq!(
-			persisted_state
-				.voting_oracle()
-				.voting_target(persisted_state.best_beefy_block(), 13),
-			Some(12)
-		);
-
-		// verify state also saved to db
-		let version: u32 = load_decode(&*backend, VERSION_KEY).unwrap().unwrap();
-		assert_eq!(version, CURRENT_VERSION);
-		let state = load_decode::<_, PersistedState<Block>>(&*backend, WORKER_STATE)
-			.unwrap()
-			.unwrap();
-		assert_eq!(state, persisted_state);
-	}
-
-	#[test]
-	fn should_initialize_voter_at_latest_finalized() {
-		let keys = &[Keyring::Alice];
-		let validator_set = ValidatorSet::new(make_beefy_ids(keys), 0).unwrap();
-		let mut net = BeefyTestNet::new(1);
-		let backend = net.peer(0).client().as_backend();
-
-		// push 15 blocks with `AuthorityChange` digests every 10 blocks
-		net.generate_blocks_and_sync(15, 10, &validator_set, false);
-
-		let mut finality = net.peer(0).client().as_client().finality_notification_stream().fuse();
-
-		// finalize 13 without justifications
-		let hashof13 =
-			backend.blockchain().expect_block_hash_from_id(&BlockId::Number(13)).unwrap();
-		net.peer(0).client().as_client().finalize_block(hashof13, None).unwrap();
-
-		// import/append BEEFY justification for block 12
-		let commitment = Commitment {
-			payload: Payload::from_single_entry(known_payloads::MMR_ROOT_ID, vec![]),
-			block_number: 12,
-			validator_set_id: validator_set.id(),
-		};
-		let justif = VersionedFinalityProof::<_, Signature>::V1(SignedCommitment {
-			commitment,
-			signatures: vec![None],
-		});
-		let hashof12 =
-			backend.blockchain().expect_block_hash_from_id(&BlockId::Number(12)).unwrap();
-		backend
-			.append_justification(hashof12, (BEEFY_ENGINE_ID, justif.encode()))
-			.unwrap();
-
-		// Test initialization at last BEEFY finalized.
-
-		// load persistent state - nothing in DB, should init at last BEEFY finalized
-		let persisted_state = load_persistent_helper(&mut net, &mut finality);
-
-		// verify voter initialized with single session starting at block 12
-		assert_eq!(persisted_state.voting_oracle().sessions().len(), 1);
-		let rounds = persisted_state.active_round().unwrap();
-		assert_eq!(rounds.session_start(), 12);
-		assert_eq!(rounds.validator_set_id(), validator_set.id());
-
-		// verify next vote target is 13
-		assert_eq!(persisted_state.best_beefy_block(), 12);
-		assert_eq!(persisted_state.best_grandpa_block(), 13);
-		assert_eq!(
-			persisted_state
-				.voting_oracle()
-				.voting_target(persisted_state.best_beefy_block(), 13),
-			Some(13)
-		);
-
-		// verify state also saved to db
-		let version: u32 = load_decode(&*backend, VERSION_KEY).unwrap().unwrap();
-		assert_eq!(version, CURRENT_VERSION);
-		let state = load_decode::<_, PersistedState<Block>>(&*backend, WORKER_STATE)
-			.unwrap()
-			.unwrap();
-		assert_eq!(state, persisted_state);
+		// full `PersistedState` load is tested in `tests.rs`.
 	}
 }
